perf(avm2-gc): tier-2 collector cost — arena bitmap, epoch marks, lazy sweep
The adaptive watermark (tier 1) fixed collect FREQUENCY; this cuts the
per-object cost of a collect itself. RWK gameplay collect pause 73 -> ~23 ms
native (600-tick plan_k TAS, per-collect phase timer):

  phase      HEAD   +bitmap  +epochs  +lazy sweep
  snap       36.5    19.8      0.02      0.02
  trace      11.2    10.1     11.5      15.5
  sweep      21.8    22.0     24.4       2.6   (rest amortized ~2-3 ms/tick)
  total      73.0    55.5     40.0      ~23

Lever 1 — arena membership bitmap. The conservative ext scan's "does a census
object start here?" test was a binary search over a sorted snapshot rebuilt
(walk + qsort) every collect: half the pause at a 284k census. Now 1 bit per
o1heap alignment cell, set on enroll and cleared on sweep, so membership is
O(1) with no per-collect work. New heap_arena_base/span/alignment() accessors;
the snapshot path stays as the HEAP_PASSTHROUGH (sanitizer) fallback.

Lever 2 — epoch marks. gc_mark widened uint8 -> uint32 INTO EXISTING PADDING
(sizeof and every offset unchanged, verified on native and wasm32): bit 0 =
pinned, bits 1.. = mark epoch. Bumping the epoch whitens the census in O(1),
so the clear walk is gone; pins moved to their own array to seed the worklist.
avm2_gc_is_marked() replaces the one external gc_mark reader.

Lever 3 — lazy sweep. Marking stays atomic between ticks; the sweep runs from
a resumable cursor at AVM2_GC_SWEEP_BUDGET entries per tick (default 25000
~= 2 ms; 0 = eager, forced under AVM2_GC_STRESS). Objects born during a sweep
enroll in a nursery and are born black, so the cursor never meets them; the
nursery splices back when the cursor finishes and no mark starts before that.
Below 1/8 arena headroom the sweep runs eagerly — rwic lives at ~96% of its
arena, where an allocation failure is fatal and a long pause is not.

Rig (Windows real-GPU Chrome, interleaved rounds, all-frames mean / frames
over 250 ms): rwk 103.5 -> 81.6 ms with 30 stalls of 463-820 ms -> ZERO;
rwp 64.8 -> 54.8 ms, 14 -> 0; seedling 26.0 -> 22.7 ms, 1 -> 0. Note cpu p50
RISES (72.7 -> 83.1 on rwk) by design: the sweep moved out of the freeze into
the following ticks. Judge on stalls + all-frames mean, not p50.

Verification: traces byte-identical to HEAD at 600 and 1200 ticks across
normal/stress/GC=0; 600 CPU-dump frames byte-identical; collect cadence,
census size, swept counts and total-alloc identical; maxRSS flat; the three
avm2_gc_* regression tests pass in all three GC modes.

+++ SWFModernRuntime/include/avm2/avm2_object.h
@@ -83,7 +83,14 @@ struct Avm2Object
 	// scanned); a non-zero value means "conservatively scan this many bytes of
 	// native_ext for embedded object pointers".
 	Avm2Object* gc_next;
-	uint8_t gc_mark;          // 0 white, 1 marked, 2 pinned (never swept)
+	// bit 0 = pinned (immortal: class objects / prototypes / XML nodes);
+	// bits 1.. = MARK EPOCH — the object is marked iff (gc_mark >> 1) equals
+	// the collector's current epoch, which increments once per collect. That
+	// makes "reset every object to white" O(1) instead of a walk of the whole
+	// census (tier-2 lever 2; see avm2_gc.c). Widened from uint8_t into the
+	// padding that already sat between it and native_ext_size — sizeof and
+	// every field offset are unchanged (static_assert in avm2_gc.c).
+	uint32_t gc_mark;
 	uint32_t native_ext_size;
 
 	// AVM2_OBJ_FUNCTION payload: a bound method closure.
